Add temperature and humidity display demo
This commit adds all the required files for the demo, including the DHT
library.  The code reads the current DHT values, sends them over the
serial port, which is then displayed on the Grove LCD RGB display.  The
sensor's arduino deals with the DHT module, and the baseboard drives the
LCD display.

Signed-off-by: Sophie Haynes <sophie.haynes@linaro.org>
Signed-off-by: Grant Likely <grant.likely@linaro.org>